refactor(anni): reorganize components
