+++ annotation-interface/components/common/SelectionPopover.tsx
