+++ annotation-interface/pages/_app.tsx
@@ -2,7 +2,7 @@ import '../styles/globals.css';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
 
-import Layout from '../components/Layout';
+import Layout from '../components/common/Layout';
 
 function AnnotationInterface({ Component, pageProps }: AppProps) {
     return (
